requirements.prod.txt
Flask==2.3.2
openai==0.27.7
pgvector==0.1.8
python-dotenv==1.0.0
tiktoken==0.4.0
urllib3==2.0.2
gunicorn==20.1.0
langchain==0.0.198
langchainplus-sdk==0.0.8
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-c==3.1.9